Finished APDevice table division & server optimizations

+++ allshoppings-services/src/main/java/mobi/allshoppings/dao/AuditLogDAO.java
@@ -12,5 +12,5 @@ import mobi.allshoppings.tools.Range;
 public interface AuditLogDAO extends GenericDAO<AuditLog> {
 
 	Key createKey() throws ASException;
-	List<AuditLog> getUsingUserAndTypeAndRange(String userId, Integer eventType, Range range, String order, Map<String, String> attributes, boolean detachable) throws ASException;
+	List<AuditLog> getUsingUserAndTypeAndRange(String userId, byte eventType, Range range, String order, Map<String, String> attributes, boolean detachable) throws ASException;
 }
